add functions on module
